Move v33 handler to keyboard.c.
