+++ drivers/keyboard.h
@@ -125,4 +125,6 @@ int get_scan_code2(uint8_t *sc);
 
 void sc_sm_update(uint8_t in);
 
+void v33_handler(uint32_t vn, uint32_t err_code);
+
 #endif
